Add files via upload
The Fibonacci Series  is a type series where each number is the sum of the two that precede it. It starts from 0 and 1 usually. The Fibonacci sequence is given by 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, and so on. Here is the python code to generate the Fibonacci Series by using the while loop. It's get the input from the user and produce the output. For example if the user gives the input 5 it gives the output like [0, 1, 1, 2, 3].